Time: 237 ms (30.20%) | Memory: 147.9 MB (20.62%) - LeetSync

+++ 2455-node-with-highest-edge-score/node-with-highest-edge-score.cpp
@@ -2,13 +2,13 @@
 class Solution {
 public:
     int edgeScore(vector<int>& edges) {
-        unordered_map<ll,ll> sum;
+        unordered_map<int,ll> sum;
         ll maxSum = 0;
         for(int i = 0; i<edges.size(); i++){
             sum[edges[i]] += i;
             maxSum = max(maxSum, sum[edges[i]]);
         }
-        ll node = edges.size();
+        int  node = edges.size();
 
         for(auto &x: sum){
             if(x.second == maxSum){
@@ -16,7 +16,7 @@ public:
             }
         }
 
-        return node;
+        return (int) node;
